+++ cmp/grid/columns/Column.js
@@ -95,10 +95,6 @@ export class Column {
 
     /** @member {boolean} */
     hidden;
-    /** @member {boolean} */
-    hideWhenGrouped;
-    /** @member {boolean} */
-    showWhenUngrouped;
     /** @member {(boolean|number)} */
     flex;
     /** @member {number} */
@@ -229,8 +225,6 @@ export class Column {
      *      class names to functions determining if they should be added or removed from the cell.
      *      See Ag-Grid docs on "cell styles" for details.
      * @param {boolean} [c.hidden] - true to suppress default display of the column.
-     * @param {boolean} [c.hideWhenGrouped] - true to hide column when grouped.
-     * @param {boolean} [c.showWhenUngrouped] - true to show column when ungrouped.
      * @param {string} [c.align] - horizontal alignment of cell contents.
      *      Valid values are:  'left' (default), 'right' or 'center'.
      * @param {number} [c.width] - default width in pixels.
@@ -281,8 +275,8 @@ export class Column {
      * @param {string} [c.chooserName] - name to display within the column chooser component.
      *      Defaults to `displayName`, can be longer / less abbreviated than `headerName` might be.
      * @param {string} [c.chooserGroup] - group name to display within the column chooser
-     *     component.
-     *      Chooser will automatically group its "available columns" grid if any cols provide.
+     *     component. Chooser will automatically group its "available columns" grid if any cols
+     *     provide.
      * @param {string} [c.chooserDescription] - additional descriptive text to display within the
      *      column chooser. Appears when the column is selected within the chooser UI.
      * @param {boolean} [c.excludeFromChooser] - true to hide the column from the column chooser
@@ -344,8 +338,6 @@ export class Column {
         cellClass,
         cellClassRules,
         hidden,
-        hideWhenGrouped,
-        showWhenUngrouped,
         align,
         width,
         minWidth,
@@ -425,8 +417,6 @@ export class Column {
         this.align = align;
 
         this.hidden = withDefault(hidden, false);
-        this.hideWhenGrouped = withDefault(hideWhenGrouped, true);
-        this.showWhenUngrouped = withDefault(showWhenUngrouped, false);
         warnIf(rest.hide, `Column ${this.colId} configured with {hide: true} - use "hidden" instead.`);
 
         warnIf(
@@ -477,10 +467,7 @@ export class Column {
         this.chooserName = chooserName || this.displayName;
         this.chooserGroup = chooserGroup;
         this.chooserDescription = chooserDescription;
-        this.excludeFromChooser = withDefault(showWhenUngrouped && hidden ? false : excludeFromChooser, false);
-        warnIf(showWhenUngrouped && hidden && excludeFromChooser,
-            `Column "${this.colId}" cannot be configured with {hidden: true, excludeFromChooser: true} AND {showWhenUngrouped: true}. ExcludeFromChooser will be ignored.`
-        );
+        this.excludeFromChooser = withDefault(excludeFromChooser, false);
 
         // ExportName must be non-empty string. Default to headerName if unspecified (it supports
         // the function form of headerName) and fallback to colId. Note GridExportService can
@@ -515,6 +502,16 @@ export class Column {
         // Warn if using the ag-Grid valueSetter or valueGetter and recommend using our callbacks
         warnIf(this.agOptions.valueSetter, `Column '${this.colId}' uses valueSetter through agOptions. Remove and use custom setValueFn if needed.`);
         warnIf(this.agOptions.valueGetter, `Column '${this.colId}' uses valueGetter through agOptions. Remove and use custom getValueFn if needed.`);
+    }
+
+    /** @return {boolean} - true if column should always remain visible */
+    get lockVisible() {
+        return !this.hideable || !this.gridModel.colChooserModel || XH.isMobileApp;
+    }
+
+    /** @return {boolean} - true if column is initialized as hidden and cannot be made visible via ChooserModel. */
+    get lockHidden() {
+        return this.hidden && (this.excludeFromChooser || !this.gridModel.colChooserModel);
     }
 
     /**
@@ -553,7 +550,7 @@ export class Column {
                 suppressMovable: !this.movable,
                 lockPinned: !gridModel.enableColumnPinning || XH.isMobileApp,
                 pinned: this.pinned,
-                lockVisible: !this.hideable || !gridModel.colChooserModel || XH.isMobileApp,
+                lockVisible: this.lockVisible,
                 headerComponentParams: {gridModel, xhColumn: this},
                 suppressColumnsToolPanel: this.excludeFromChooser,
                 suppressFiltersToolPanel: this.excludeFromChooser,